clean: Moved Bool types outside interface.analog, for use by other interfaces. Rename restricted_type to Identifier
